ICUI-21: moved ProtectedRoute into its own file
**Problem:**

**Fix:**

**Testing:**
